+++ client/pages/_app.tsx
@@ -1,9 +1,9 @@
 import '../styles/globals.css';
 import type { AppProps } from 'next/app';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import DefaultLayout from '../components/layout/defalutLayout/DefaultLayout';
+import DefaultLayout from '../components/container/defalutLayout/DefaultLayout';
 import { ReactElement } from 'react';
-import { NextPageWithLayout } from '../components/layout/defalutLayout/defaultLayout';
+import { NextPageWithLayout } from '../components/container/defalutLayout/defaultLayout';
 import { CookiesProvider } from 'react-cookie';
 
 type AppPropsWithLayout = AppProps & {
