+++ client/pages/index.tsx
@@ -5,8 +5,8 @@ import SearchOutlinedIcon from '@mui/icons-material/SearchOutlined';
 import Button from '@mui/material/Button';
 import MainSlogan from '../components/molecules/slogan/mainSlogan/MainSlogan';
 import Footer from '../components/molecules/footer/Footer';
-import LayoutWithFooter from '../components/layout/layoutWithFooter/LayoutWithFooter';
-import { NextPageWithLayout } from '../components/layout/defalutLayout/defaultLayout';
+import LayoutWithFooter from '../components/container/layoutWithFooter/LayoutWithFooter';
+import { NextPageWithLayout } from '../components/container/defalutLayout/defaultLayout';
 
 const MainPage: NextPageWithLayout = () => {
   const [nearBy, setNearBy] = useState('');
